% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{contigs2bins}
\alias{contigs2bins}
\title{Get the contigs associated with each MAG.}
\usage{
contigs2bins(bin_dir)
}
\arguments{
\item{bin_dir}{Path to directory containing nucleotide fasta files of bins.}
}
\value{
A data frame with all bins and their associated contigs.
}
\description{
This function takes a directory containing the nucleotide fasta files of all of your MAGs and
outputs a data frame connecting each contig to a bin.
}
