% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{drep_members}
\alias{drep_members}
\title{Find which original bins are represented by dereplicated ones (dRep)}
\usage{
drep_members(drep_dir)
}
\arguments{
\item{drep_dir}{Path to dRep output directory}
}
\value{
A data frame with all bins not in the dereplicated genomes linked to the ones chosen by dRep.
}
\description{
This function takes the dRep output directory and outputs a data frame linking bins not in the
dereplicated genomes to those in the dereplicated genomes.
}
