% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{frac_abund}
\alias{frac_abund}
\title{Calculate reads mapped per contig}
\usage{
frac_abund(
  coverm_out,
  bin_dir,
  gtdbtk_dir,
  flagstat_dir,
  min_reads = 10,
  cov_cut,
  prop = T,
  per_ml,
  rank_ids = TRUE,
  tax_rank = "bin",
  gene = F
)
}
\arguments{
\item{coverm_out}{Text file of the output of CoverM. CoverM command should use the methods length, count, and covered_bases.}

\item{bin_dir}{Path to directory containing nucleotide fasta files of bins. Only the nucleotide files containing
contigs should be present. No other files should be present i.e. faa, gbk, nucleotide files of genes, etc.}

\item{gtdbtk_dir}{Path to GTDB-tk output directory}

\item{flagstat_dir}{Path to directory with files containing the raw output of samtools flagstat for each sample.
File names should end in _flagstat followed by .csv or .tsv or .txt. Everything before "\emph{flagstat" will be used as the sample names.
Do not include another "}" after "_flagstat".}

\item{min_reads}{Minimum number of reads mapped to a contig to be considered present in a metagenome. Default is 10 reads.}

\item{cov_cut}{Minimum breadth of coverage for a contig to be considered there. The idea here is that X\% of the contig
needs to be covered by reads within a sample to be considered present. Default is to not perform coverage filtering.}

\item{prop}{TRUE or FALSE indicating if you want the output to be on a 1-100 percent scale. FALSE will output the
fractional abundance as defined by FRAP. The default is TRUE.}

\item{per_ml}{Data frame containing the sample name in the 1st column and the number of bacteria per mL of each sample.
Make sure to include all samples if you have replicate for the metagenomes but are using one count value then you will need
multiple rows for each replicate with the appropriate names.}

\item{rank_ids}{Determines whether or not taxonomic information will retain the prefix indicating the taxonomic level.
Default is TRUE which retains the prefixes. FALSE will remove the prefixes.}

\item{tax_rank}{The taxonomic rank at which you would like the fractional abundance to be calculated.
Ranks that can be used are "bin", "domain", "phylum", "class", "order", "family", "genus", "species".
The input should be a character string. Default is "bin".}

\item{gene}{T or F indicating if mapping was done with genes from prodigal (T) or on contigs (F). Default value is F.}
}
\value{
The output will be a data frame containing the fractional abundance of each taxonomic rank chosen.
}
\description{
The functions has 4 required parameters: CoverM output file a directory containing a separate
nucleotide fasta file per bin, a directory containing the output of GTDB-tk, and a directory
containing the outputs of samtools flagstat for each sample you mapped to your bins.
}
