% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{makecolors}
\alias{makecolors}
\title{Generate a nice color pallette.}
\usage{
makecolors(color_ls, num_color_ls)
}
\arguments{
\item{color_ls}{A character list of the R colors you want to serve as the basis for your pallete.
These should be pretty distinct colors.}

\item{num_color_ls}{Each number in this numeric list should correspond to the number times you want
to iterate each color in color_ls.
#' @return The output will be a vector containing unique colors.}
}
\description{
The functions takes a list of colors and the number of times you want those colors to be re-used and
makes a larger pallete containing distinct colors.
}
