% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{parse_coverm}
\alias{parse_coverm}
\title{Parse a CoverM output}
\usage{
parse_coverm(coverm_tsv)
}
\arguments{
\item{coverm_tsv}{Path to tsv file containing the CoverM output.}
}
\value{
The output will be a dataframe containing the contig names and number of reads mapped per output file in your directory.
}
\description{
This function is intended for use in calculating fractional abundances. It parses a CoverM output for downstream
analysis the following CoverM command will generate the expected input:
\verb{coverm contig [dir_of_bams]/*.bam --methods length count covered_bases > coverm_output.tsv}.
}
