% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_nmds}
\alias{plot_nmds}
\title{Make NMDS plots.}
\usage{
plot_nmds(taxa, groups, dis_ind = "bray", p_color, p_shape)
}
\arguments{
\item{taxa}{Dataframe containing the taxa of interest as rows and the abundances per sample as
numeric columns. Additional character columns with taxonomic information can be left in as is
the standard output of frac_abund().}

\item{num_color_ls}{Each number in this numeric list should correspond to the number times you want
to iterate each color in color_ls.}
}
\value{
The output will be a vector containing unique colors.
}
\description{
The functions takes a list of colors and the number of times you want those colors to be re-used and
makes a larger pallete containing distinct colors.
}
