% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_gtdbtk}
\alias{read_gtdbtk}
\title{Read in taxonomic information generated by GTDB-tk}
\usage{
read_gtdbtk(gtdbtk_dir, rank_ids = T)
}
\arguments{
\item{gtdbtk_dir}{Path to GTDB-tk output directory}

\item{rank_ids}{Determines whether or not taxonomic information will retain the prefix indicating the taxonomic level.
The default is TRUE which retains the prefixes. FALSE will remove the prefixes}
}
\value{
A data frame with all MAG taxonomies. Empty fields are changed to
"Unclassified" followed by the lowest classifiable taxonomy.
}
\description{
This function loads in taxonomic information created by GTDB-tk.
It aggregates the archaeal and bacterial data, populates empty fields,
and creates separate fields for each taxonomic rank.
}
