% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{reads_meta}
\alias{reads_meta}
\title{Get total read counts per metagenome.}
\usage{
reads_meta(flagstat_dir)
}
\arguments{
\item{flagstat_dir}{Path to samtools directory. Directory should have the files containing the raw output of
samtools flagstat for each of your samples. File names should end in _flagstat followed by .csv or .tsv or .txt.
Everything before "\emph{flagstat" will be used as the sample names. Do not include another "}" after "_flagstat".}
}
\value{
The output will be a data frame with 1 row containing all of your reads per metagenome.
}
\description{
This function loads in all of your samtools flagstat output files and
creates a single data frame with all sample reads per metagenome.
}
