% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{contigs2bins}
\alias{contigs2bins}
\title{Get the contigs associated with each MAG.}
\usage{
contigs2bins(bin_dir, suffix = "fna")
}
\arguments{
\item{bin_dir}{Path to directory containing nucleotide fasta files of bins.}

\item{suffix}{Ending of fasta files containing contigs in nucleotides. This just changes the string that list files uses as the pattern.
Default is "fna".}
}
\value{
A data frame with all bins and their associated contigs.
}
\description{
This function takes a directory containing the nucleotide fasta files of all of your MAGs and
outputs a data frame connecting each contig to a bin.
}
